REF: move alignment functions to DataFrame/Series methods
