+++ pandas/core/computation/expressions.py
@@ -17,8 +17,8 @@ from pandas._config import get_option
 
 from pandas.util._exceptions import find_stack_level
 
+from pandas.core import roperator
 from pandas.core.computation.check import NUMEXPR_INSTALLED
-from pandas.core.ops import roperator
 
 if NUMEXPR_INSTALLED:
     import numexpr as ne
