+++ pandas/core/ops/missing.py
@@ -33,7 +33,7 @@ from pandas.core.dtypes.common import (
     is_scalar,
 )
 
-from pandas.core.ops import roperator
+from pandas.core import roperator
 
 
 def _fill_zeros(result, x, y):
