+++ pandas/tests/frame/test_arithmetic.py
@@ -1800,7 +1800,7 @@ class TestFrameArithmeticUnsorted:
         columns = ["X", "Y", "Z"]
         df = DataFrame(np.random.randn(3, 3), index=index, columns=columns)
 
-        align = pd.core.ops.align_method_FRAME
+        align = DataFrame._align_for_op
 
         expected = DataFrame({"X": val, "Y": val, "Z": val}, index=df.index)
         tm.assert_frame_equal(align(df, val, "index")[1], expected)
@@ -1816,7 +1816,7 @@ class TestFrameArithmeticUnsorted:
         columns = ["X", "Y", "Z"]
         df = DataFrame(np.random.randn(3, 3), index=index, columns=columns)
 
-        align = pd.core.ops.align_method_FRAME
+        align = DataFrame._align_for_op
         # length mismatch
         msg = "Unable to coerce to Series, length must be 3: given 2"
         with pytest.raises(ValueError, match=msg):
@@ -1830,7 +1830,7 @@ class TestFrameArithmeticUnsorted:
         columns = ["X", "Y", "Z"]
         df = DataFrame(np.random.randn(3, 3), index=index, columns=columns)
 
-        align = pd.core.ops.align_method_FRAME
+        align = DataFrame._align_for_op
         val = np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
         tm.assert_frame_equal(
             align(df, val, "index")[1],
